backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.26.0
google-genai>=1.0.0
slowapi>=0.1.9
pymongo>=4.7.0
motor>=3.4.0
